Changed sendDestroyMsg's access level to private, since it should only be called by Observable's destructor. This ensures that sendDestroyMsg is sent only when Observable's are destroyed, and furthermore that it is only sent once during the destruction process.
Added a detach() during sendDestroyMsg to ensure that each Observer is
removed from the list of observers once it has been notified of the
Observable's destruction. This allows Observer's not to have to worry
about detach()'ing themselves, and ensures that Observer's do not
receive any new notifications once they have been informed of the
Observable's destruction.